+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## Unreleased
+- Allows using event names that exist on `Object.prototype`.
+
 ## 0.1.337 (Oct 25, 2013)
 - First release. The show must go on!